ViewerImpl.cpp: Fix signed/unsigned comparison warnings by changing loop counters from int to size_t

+++ src/libvgcode/src/ViewerImpl.cpp
@@ -186,7 +186,7 @@ static Mat4x4 inverse(const Mat4x4& m)
     det = 1.0f / det;
 
     std::array<float, 16> ret = {};
-    for (int i = 0; i < 16; ++i) {
+    for (size_t i = 0; i < 16; ++i) {
         ret[i] = inv[i] * det;
     }
 
@@ -1888,7 +1888,7 @@ void ViewerImpl::render_segments(const Mat4x4& view_matrix, const Mat4x4& projec
     }
 #else
     std::array<int, 4> curr_bound_texture = { 0, 0, 0, 0 };
-    for (int i = 0; i < curr_bound_texture.size(); ++i) {
+    for (size_t i = 0; i < curr_bound_texture.size(); ++i) {
         glsafe(glActiveTexture(GL_TEXTURE0 + i));
         glsafe(glGetIntegerv(GL_TEXTURE_BINDING_BUFFER, &curr_bound_texture[i]));
         //assert(curr_bound_texture[i] == 0);
@@ -1917,7 +1917,7 @@ void ViewerImpl::render_segments(const Mat4x4& view_matrix, const Mat4x4& projec
 #ifdef ENABLE_OPENGL_ES
     glsafe(glBindTexture(GL_TEXTURE_2D, curr_bound_texture));
 #else
-    for (int i = 0; i < curr_bound_texture.size(); ++i) {
+    for (size_t i = 0; i < curr_bound_texture.size(); ++i) {
         glsafe(glActiveTexture(GL_TEXTURE0 + i));
         glsafe(glBindTexture(GL_TEXTURE_BUFFER, curr_bound_texture[i]));
     }
@@ -1975,7 +1975,7 @@ void ViewerImpl::render_options(const Mat4x4& view_matrix, const Mat4x4& project
     }
 #else
     std::array<int, 4> curr_bound_texture = { 0, 0, 0, 0 };
-    for (int i = 0; i < curr_bound_texture.size(); ++i) {
+    for (size_t i = 0; i < curr_bound_texture.size(); ++i) {
         glsafe(glActiveTexture(GL_TEXTURE0 + i));
         glsafe(glGetIntegerv(GL_TEXTURE_BINDING_BUFFER, &curr_bound_texture[i]));
         //assert(curr_bound_texture[i] == 0);
@@ -2004,7 +2004,7 @@ void ViewerImpl::render_options(const Mat4x4& view_matrix, const Mat4x4& project
 #ifdef ENABLE_OPENGL_ES
     glsafe(glBindTexture(GL_TEXTURE_2D, curr_bound_texture));
 #else
-    for (int i = 0; i < curr_bound_texture.size(); ++i) {
+    for (size_t i = 0; i < curr_bound_texture.size(); ++i) {
         glsafe(glActiveTexture(GL_TEXTURE0 + i));
         glsafe(glBindTexture(GL_TEXTURE_BUFFER, curr_bound_texture[i]));
     }
